Move MeshAnimation to a separate class Drawable.
The move makes it easier to refactor. The renaming helps understand the function better. At a later stage, the Unit class can simply inherit from Drawable.
