+++ engine/CMakeLists.txt
@@ -203,6 +203,8 @@ SET(LIBCMD_SOURCES
     src/cmd/unit_util_generic.cpp
     src/cmd/unit_xml.cpp
     src/cmd/weapon_xml.cpp
+
+    src/cmd/drawable.cpp
 )
 
 SET(LIBCMDSERVER_SOURCES
