+++ engine/src/cmd/unit_generic.h
@@ -28,6 +28,8 @@
 
 #ifndef _UNIT_H_
 #define _UNIT_H_
+#include "drawable.h"
+
 #ifdef VS_DEBUG
 #define CONTAINER_DEBUG
 #endif
@@ -59,6 +61,7 @@ void UncheckUnit( class Unit*un );
 #include "collide_map.h"
 #include "SharedPool.h"
 
+
 extern char * GetUnitDir( const char *filename );
 extern float capship_size;
 
@@ -85,7 +88,6 @@ class Box;
 class StarSystem;
 struct colTrees;
 class Pilot;
-class MeshAnimation;
 
 /**
  * Currently the only inheriting function is planet
@@ -1551,7 +1553,8 @@ public:
     **** ANIMATION STUFF                                                       ***
     **************************************************************************************
     */
-        MeshAnimation *pMeshAnimation;	
+    // TODO: std::unique_ptr<Drawable> pMeshAnimation;
+    Drawable *pMeshAnimation;
 };
 
 Unit * findUnitInStarsystem( const void *unitDoNotDereference );
@@ -1616,75 +1619,6 @@ extern void ClearDowngradeMap();
  **** MESH ANIMATION STUFF                                                       ***
  **************************************************************************************
  */
-
-class MeshAnimation
-{
-protected:
-    std::vector< std::vector<Mesh *> *> vecAnimations;
-    std::vector< string > vecAnimationNames;
-
-    bool animatedMesh;
-    unsigned int activeAnimation;
-    double timeperframe;
-    bool   done;
-	unsigned int activeMesh;
-	unsigned int nextactiveMesh;
-	bool infiniteLoop;
-	unsigned int    loopCount;
-
-	string uniqueUnitName;
-    Unit *unitDst;
-
-public:
-    double curtime;
-
-    static unsigned int unitCount;
-
-    static std::map< string, Unit * > Units;
-
-    MeshAnimation(Unit* _unitDst);
-
-    bool Init(const char *filename, int faction, Flightgroup *flightgrp = NULL, const char *animationExt = NULL);
-
-    static void UpdateFrames();
-
-    void AnimationStep();
-
-    void clear();
-
-    ~MeshAnimation() { clear(); }
-
-    string getAnimationName(unsigned int animationNumber) const;
-
-    unsigned int getAnimationNumber(const char *name) const;
-
-    void ChangeAnimation( const char *name );
-
-    void ChangeAnimation( unsigned int AnimNumber );
-
-    //set how_many_times to 0 for continuous loop animation
-    void StartAnimation( unsigned int how_many_times = 0, int numAnimation = 0 );
-
-    void StopAnimation();
-
-    bool isAnimatedMesh() const;
-
-    bool animationRuns() const;
-
-    unsigned int numAnimations();
-
-    bool isContinuousLoop() const;
-
-    void addAnimation( std::vector<Mesh *> *meshes, const char* name );
-
-    double framesPerSecond() const;
-
-    double timePerFrame() const;
-
-    void ToggleAnimatedMesh( bool on );
-
-    void SetAniSpeed( float speed );
-};
 
 
 #endif
