+++ engine/src/star_system.cpp
@@ -39,6 +39,7 @@
 #include "vsfilesystem.h"
 #include "cmd/unit_find.h"
 #include "gfx/occlusion.h"
+#include "drawable.h"
 
 #include "options.h"
 
@@ -376,7 +377,7 @@ void GameStarSystem::Draw( bool DrawCockpit )
     ConditionalCursorDraw( false );
     if (DrawCockpit)
         _Universe->AccessCockpit()->Draw();
-    MeshAnimation::UpdateFrames();
+    Drawable::UpdateFrames();
     
     // And now we're done with the occluder set
     Occlusion::end();
